+++ lib/PhpParser/Internal/PrintableNewAnonClassNode.php
@@ -41,7 +41,6 @@ class PrintableNewAnonClassNode extends Expr
         assert($class instanceof Node\Stmt\Class_);
         // We don't assert that $class->name is null here, to allow consumers to assign unique names
         // to anonymous classes for their own purposes. We simplify ignore the name here.
-        assert($class->name === null);
         return new self(
             $newNode->args, $class->extends, $class->implements,
             $class->stmts, $newNode->getAttributes()