+++ lib/PhpParser/Node/Stmt/Property.php
@@ -3,6 +3,9 @@
 namespace PhpParser\Node\Stmt;
 
 use PhpParser\Node;
+use PhpParser\Node\Identifier;
+use PhpParser\Node\Name;
+use PhpParser\Node\NullableType;
 
 class Property extends Node\Stmt
 {
@@ -10,22 +13,26 @@ class Property extends Node\Stmt
     public $flags;
     /** @var PropertyProperty[] Properties */
     public $props;
+    /** @var null|Identifier|Name|NullableType Type declaration */
+    public $type;
 
     /**
      * Constructs a class property list node.
      *
-     * @param int                $flags      Modifiers
-     * @param PropertyProperty[] $props      Properties
-     * @param array              $attributes Additional attributes
+     * @param int                                      $flags      Modifiers
+     * @param PropertyProperty[]                       $props      Properties
+     * @param array                                    $attributes Additional attributes
+     * @param null|string|Identifier|Name|NullableType $type       Type declaration
      */
-    public function __construct(int $flags, array $props, array $attributes = []) {
+    public function __construct(int $flags, array $props, array $attributes = [], $type = null) {
         parent::__construct($attributes);
         $this->flags = $flags;
         $this->props = $props;
+        $this->type = \is_string($type) ? new Identifier($type) : $type;
     }
 
     public function getSubNodeNames() : array {
-        return ['flags', 'props'];
+        return ['flags', 'type', 'props'];
     }
 
     /**
@@ -64,7 +71,7 @@ class Property extends Node\Stmt
     public function isStatic() : bool {
         return (bool) ($this->flags & Class_::MODIFIER_STATIC);
     }
-    
+
     public function getType() : string {
         return 'Stmt_Property';
     }